refactor(gui): extract input and select components
Why:
- reuse input and select widgets across windows

What:
- move TextInput/SelectInput/InputKind into shared inputs module
- update config editor and subtitle editor to use new module
- keep shared input key bindings exported from components

Tests:
- cargo fmt --all
- cargo check --all-targets
- cargo clippy --all-targets -- -D warnings
