+++ crates/subtitle-fast/src/gui/components/config_editor.rs
@@ -1,23 +1,17 @@
 use std::env;
 use std::fs;
-use std::ops::Range;
 use std::path::PathBuf;
-use std::time::{Duration, Instant};
 
 use gpui::Styled;
 use gpui::prelude::*;
 use gpui::{
-    AnyElement, App, Bounds, ClipboardItem, Context, CursorStyle, DispatchPhase, Element,
-    ElementId, ElementInputHandler, Entity, EntityInputHandler, FocusHandle, Focusable,
-    GlobalElementId, InteractiveElement, KeyBinding, LayoutId, MouseButton, MouseDownEvent,
-    MouseMoveEvent, MouseUpEvent, PaintQuad, Pixels, Point, PromptButton, PromptLevel, Render,
-    ScrollHandle, ShapedLine, SharedString, Style, Subscription, TextRun, UTF16Selection,
-    UnderlineStyle, Window, WindowBounds, WindowOptions, actions, div, fill, hsla, point, px,
-    relative, rgb, rgba, size,
+    AnyElement, App, Bounds, Context, DispatchPhase, Entity, MouseButton, MouseDownEvent, Pixels,
+    PromptButton, PromptLevel, Render, ScrollHandle, SharedString, Subscription, Window,
+    WindowBounds, WindowOptions, div, hsla, point, px, rgb, size,
 };
 
 use crate::gui::components::Titlebar;
-use crate::gui::icons::{Icon, icon_sm};
+use crate::gui::components::inputs::{InputKind, SelectInput, SelectOption, TextInput};
 use crate::gui::menus;
 use crate::settings::{
     self, DecoderFileConfig, DetectionFileConfig, FileConfig, OcrFileConfig, OutputFileConfig,
@@ -28,33 +22,9 @@ use subtitle_fast_decoder::Configuration as DecoderConfiguration;
 use subtitle_fast_ocr::Configuration as OcrConfiguration;
 use subtitle_fast_validator::subtitle_detection::Configuration as DetectorConfiguration;
 
-actions!(
-    config_text_input,
-    [
-        Backspace,
-        Delete,
-        Left,
-        Right,
-        SelectLeft,
-        SelectRight,
-        SelectAll,
-        Home,
-        End,
-        Tab,
-        TabPrev,
-        Enter,
-        ShowCharacterPalette,
-        Paste,
-        Cut,
-        Copy,
-    ]
-);
-
 const FIELD_LABEL_WIDTH: f32 = 200.0;
-const INPUT_HEIGHT: f32 = 30.0;
 const ERROR_ROW_HEIGHT: f32 = 14.0;
 const ERROR_ROW_PADDING: f32 = 3.0;
-const CURSOR_BLINK_INTERVAL: Duration = Duration::from_millis(500);
 
 pub struct ConfigWindow {
     fields: ConfigFields,
@@ -376,7 +346,7 @@ impl ConfigWindow {
             subscriptions.push(cx.observe(&input, |this, _input, cx| {
                 this.handle_field_change(cx);
             }));
-            let focus_handle = input.read(cx).focus_handle.clone();
+            let focus_handle = input.read(cx).focus_handle();
             subscriptions.push(cx.on_blur(&focus_handle, window, |this, _window, cx| {
                 this.handle_autosave(cx);
             }));
@@ -495,7 +465,7 @@ impl ConfigWindow {
             self.fields.decoder_backend.clone(),
             self.fields.ocr_backend.clone(),
         ];
-        selects.into_iter().any(|select| select.read(cx).open)
+        selects.into_iter().any(|select| select.read(cx).is_open())
     }
 
     fn select_popup(&mut self, window: &Window, cx: &mut Context<Self>) -> Option<AnyElement> {
@@ -509,12 +479,12 @@ impl ConfigWindow {
         for select_entity in selects {
             let (bounds, options, selected) = {
                 let select = select_entity.read(cx);
-                if !select.open {
+                if !select.is_open() {
                     continue;
                 }
-                let bounds = select.button_bounds?;
-                let options = select.options.clone();
-                let selected = select.selected;
+                let bounds = select.button_bounds()?;
+                let options = select.options_snapshot();
+                let selected = select.selected_index();
                 (bounds, options, selected)
             };
 
@@ -1038,10 +1008,10 @@ impl Render for ConfigWindow {
                         ];
                         for select in selects {
                             let select = select.read(cx);
-                            if !select.open {
+                            if !select.is_open() {
                                 continue;
                             }
-                            if let Some(bounds) = select.button_bounds
+                            if let Some(bounds) = select.button_bounds()
                                 && bounds.contains(&position)
                             {
                                 inside = true;
@@ -1327,131 +1297,6 @@ fn github_ci_active() -> bool {
         .unwrap_or(false)
 }
 
-#[derive(Clone)]
-struct SelectOption {
-    label: SharedString,
-    value: SharedString,
-}
-
-impl SelectOption {
-    fn new(label: impl Into<SharedString>, value: impl Into<SharedString>) -> Self {
-        Self {
-            label: label.into(),
-            value: value.into(),
-        }
-    }
-}
-
-struct SelectInput {
-    options: Vec<SelectOption>,
-    selected: usize,
-    open: bool,
-    button_bounds: Option<Bounds<Pixels>>,
-}
-
-impl SelectInput {
-    fn new(options: Vec<SelectOption>, default_value: &str) -> Self {
-        let selected = options
-            .iter()
-            .position(|option| option.value.as_ref() == default_value)
-            .unwrap_or(0);
-        Self {
-            options,
-            selected,
-            open: false,
-            button_bounds: None,
-        }
-    }
-
-    fn set_value(&mut self, value: SharedString, cx: &mut Context<Self>) {
-        if let Some(index) = self.options.iter().position(|option| option.value == value)
-            && self.selected != index
-        {
-            self.selected = index;
-            cx.notify();
-        }
-    }
-
-    fn value(&self) -> SharedString {
-        self.options
-            .get(self.selected)
-            .map(|option| option.value.clone())
-            .unwrap_or_default()
-    }
-
-    fn toggle_open(&mut self, cx: &mut Context<Self>) {
-        self.open = !self.open;
-        cx.notify();
-    }
-
-    fn close(&mut self, cx: &mut Context<Self>) {
-        if self.open {
-            self.open = false;
-            cx.notify();
-        }
-    }
-
-    fn select(&mut self, index: usize, cx: &mut Context<Self>) {
-        if self.selected != index {
-            self.selected = index;
-        }
-        self.open = false;
-        cx.notify();
-    }
-}
-
-impl Render for SelectInput {
-    fn render(&mut self, _window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
-        let border_color = rgb(0x2b2b2b);
-        let background = rgb(0x202020);
-        let hover_bg = rgb(0x262626);
-        let text_color = hsla(0.0, 0.0, 0.9, 1.0);
-        let muted_color = hsla(0.0, 0.0, 0.65, 1.0);
-
-        let current = self
-            .options
-            .get(self.selected)
-            .map(|option| option.label.clone())
-            .unwrap_or_else(|| SharedString::from("Select"));
-
-        let selector = div()
-            .flex()
-            .items_center()
-            .justify_between()
-            .h(px(INPUT_HEIGHT))
-            .px(px(10.0))
-            .w_full()
-            .rounded(px(6.0))
-            .bg(background)
-            .border_1()
-            .border_color(border_color)
-            .text_size(px(12.0))
-            .text_color(text_color)
-            .cursor_pointer()
-            .hover(move |style| style.bg(hover_bg))
-            .child(current)
-            .child(icon_sm(Icon::ChevronDown, muted_color))
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _, _, cx| {
-                    this.toggle_open(cx);
-                }),
-            );
-
-        let handle = cx.entity();
-        let selector_wrapper = div()
-            .on_children_prepainted(move |bounds, _window, cx| {
-                let bounds = bounds.first().copied();
-                handle.update(cx, |this, _cx| {
-                    this.button_bounds = bounds;
-                });
-            })
-            .child(selector);
-
-        div().relative().w_full().child(selector_wrapper)
-    }
-}
-
 fn normalize_optional(value: &SharedString) -> Option<String> {
     let trimmed = value.trim();
     if trimmed.is_empty() {
@@ -1588,723 +1433,4 @@ fn parse_roi_fields(
 enum ButtonStyle {
     Primary,
     Secondary,
-}
-
-#[derive(Clone, Copy)]
-pub(crate) enum InputKind {
-    Integer,
-    Float,
-    Text,
-}
-
-pub(crate) struct TextInput {
-    focus_handle: FocusHandle,
-    content: SharedString,
-    placeholder: SharedString,
-    input_kind: InputKind,
-    blink_start: Instant,
-    last_focused: bool,
-    invalid: bool,
-    selected_range: Range<usize>,
-    selection_reversed: bool,
-    marked_range: Option<Range<usize>>,
-    last_layout: Option<ShapedLine>,
-    last_bounds: Option<Bounds<Pixels>>,
-    is_selecting: bool,
-}
-
-impl TextInput {
-    pub(crate) fn new(
-        cx: &mut Context<Self>,
-        placeholder: impl Into<SharedString>,
-        input_kind: InputKind,
-    ) -> Self {
-        Self {
-            focus_handle: cx.focus_handle(),
-            content: SharedString::from(""),
-            placeholder: placeholder.into(),
-            input_kind,
-            blink_start: Instant::now(),
-            last_focused: false,
-            invalid: false,
-            selected_range: 0..0,
-            selection_reversed: false,
-            marked_range: None,
-            last_layout: None,
-            last_bounds: None,
-            is_selecting: false,
-        }
-    }
-
-    pub(crate) fn set_invalid(&mut self, invalid: bool, cx: &mut Context<Self>) {
-        if self.invalid != invalid {
-            self.invalid = invalid;
-            cx.notify();
-        }
-    }
-
-    fn filter_input(&self, range: &Range<usize>, new_text: &str) -> String {
-        match self.input_kind {
-            InputKind::Integer => new_text.chars().filter(|ch| ch.is_ascii_digit()).collect(),
-            InputKind::Float => {
-                let base = format!(
-                    "{}{}",
-                    &self.content[..range.start],
-                    &self.content[range.end..]
-                );
-                let mut has_dot = base.contains('.');
-                let mut filtered = String::new();
-                for ch in new_text.chars() {
-                    if ch.is_ascii_digit() {
-                        filtered.push(ch);
-                    } else if ch == '.' && !has_dot {
-                        filtered.push(ch);
-                        has_dot = true;
-                    }
-                }
-                filtered
-            }
-            InputKind::Text => new_text
-                .chars()
-                .filter(|ch| *ch != '\n' && *ch != '\r')
-                .collect(),
-        }
-    }
-
-    pub(crate) fn set_text(&mut self, text: impl Into<SharedString>, cx: &mut Context<Self>) {
-        self.content = text.into();
-        let len = self.content.len();
-        self.selected_range = len..len;
-        self.selection_reversed = false;
-        self.marked_range = None;
-        cx.notify();
-    }
-
-    pub(crate) fn text(&self) -> SharedString {
-        self.content.clone()
-    }
-
-    fn left(&mut self, _: &Left, _: &mut Window, cx: &mut Context<Self>) {
-        if self.selected_range.is_empty() {
-            self.move_to(self.previous_boundary(self.cursor_offset()), cx);
-        } else {
-            self.move_to(self.selected_range.start, cx)
-        }
-    }
-
-    fn right(&mut self, _: &Right, _: &mut Window, cx: &mut Context<Self>) {
-        if self.selected_range.is_empty() {
-            self.move_to(self.next_boundary(self.selected_range.end), cx);
-        } else {
-            self.move_to(self.selected_range.end, cx)
-        }
-    }
-
-    fn select_left(&mut self, _: &SelectLeft, _: &mut Window, cx: &mut Context<Self>) {
-        self.select_to(self.previous_boundary(self.cursor_offset()), cx);
-    }
-
-    fn select_right(&mut self, _: &SelectRight, _: &mut Window, cx: &mut Context<Self>) {
-        self.select_to(self.next_boundary(self.cursor_offset()), cx);
-    }
-
-    fn select_all(&mut self, _: &SelectAll, _: &mut Window, cx: &mut Context<Self>) {
-        self.move_to(0, cx);
-        self.select_to(self.content.len(), cx)
-    }
-
-    fn home(&mut self, _: &Home, _: &mut Window, cx: &mut Context<Self>) {
-        self.move_to(0, cx);
-    }
-
-    fn end(&mut self, _: &End, _: &mut Window, cx: &mut Context<Self>) {
-        self.move_to(self.content.len(), cx);
-    }
-
-    fn tab(&mut self, _: &Tab, window: &mut Window, _: &mut Context<Self>) {
-        window.focus_next();
-    }
-
-    fn tab_prev(&mut self, _: &TabPrev, window: &mut Window, _: &mut Context<Self>) {
-        window.focus_prev();
-    }
-
-    fn enter(&mut self, _: &Enter, window: &mut Window, _: &mut Context<Self>) {
-        window.blur();
-    }
-
-    fn backspace(&mut self, _: &Backspace, window: &mut Window, cx: &mut Context<Self>) {
-        if self.selected_range.is_empty() {
-            self.select_to(self.previous_boundary(self.cursor_offset()), cx)
-        }
-        self.replace_text_in_range(None, "", window, cx)
-    }
-
-    fn delete(&mut self, _: &Delete, window: &mut Window, cx: &mut Context<Self>) {
-        if self.selected_range.is_empty() {
-            self.select_to(self.next_boundary(self.cursor_offset()), cx)
-        }
-        self.replace_text_in_range(None, "", window, cx)
-    }
-
-    fn on_mouse_down(
-        &mut self,
-        event: &MouseDownEvent,
-        _window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.is_selecting = true;
-
-        if event.modifiers.shift {
-            self.select_to(self.index_for_mouse_position(event.position), cx);
-        } else {
-            self.move_to(self.index_for_mouse_position(event.position), cx)
-        }
-    }
-
-    fn on_mouse_up(&mut self, _: &MouseUpEvent, _window: &mut Window, _: &mut Context<Self>) {
-        self.is_selecting = false;
-    }
-
-    fn on_mouse_move(&mut self, event: &MouseMoveEvent, _: &mut Window, cx: &mut Context<Self>) {
-        if self.is_selecting {
-            self.select_to(self.index_for_mouse_position(event.position), cx);
-        }
-    }
-
-    fn show_character_palette(
-        &mut self,
-        _: &ShowCharacterPalette,
-        window: &mut Window,
-        _: &mut Context<Self>,
-    ) {
-        window.show_character_palette();
-    }
-
-    fn paste(&mut self, _: &Paste, window: &mut Window, cx: &mut Context<Self>) {
-        if let Some(text) = cx.read_from_clipboard().and_then(|item| item.text()) {
-            self.replace_text_in_range(None, &text.replace("\n", " "), window, cx);
-        }
-    }
-
-    fn copy(&mut self, _: &Copy, _: &mut Window, cx: &mut Context<Self>) {
-        if !self.selected_range.is_empty() {
-            cx.write_to_clipboard(ClipboardItem::new_string(
-                self.content[self.selected_range.clone()].to_string(),
-            ));
-        }
-    }
-
-    fn cut(&mut self, _: &Cut, window: &mut Window, cx: &mut Context<Self>) {
-        if !self.selected_range.is_empty() {
-            cx.write_to_clipboard(ClipboardItem::new_string(
-                self.content[self.selected_range.clone()].to_string(),
-            ));
-            self.replace_text_in_range(None, "", window, cx)
-        }
-    }
-
-    fn move_to(&mut self, offset: usize, cx: &mut Context<Self>) {
-        self.selected_range = offset..offset;
-        cx.notify()
-    }
-
-    fn cursor_offset(&self) -> usize {
-        if self.selection_reversed {
-            self.selected_range.start
-        } else {
-            self.selected_range.end
-        }
-    }
-
-    fn index_for_mouse_position(&self, position: Point<Pixels>) -> usize {
-        if self.content.is_empty() {
-            return 0;
-        }
-
-        let (Some(bounds), Some(line)) = (self.last_bounds.as_ref(), self.last_layout.as_ref())
-        else {
-            return 0;
-        };
-        if position.y < bounds.top() {
-            return 0;
-        }
-        if position.y > bounds.bottom() {
-            return self.content.len();
-        }
-        line.closest_index_for_x(position.x - bounds.left())
-    }
-
-    fn select_to(&mut self, offset: usize, cx: &mut Context<Self>) {
-        if self.selection_reversed {
-            self.selected_range.start = offset
-        } else {
-            self.selected_range.end = offset
-        };
-        if self.selected_range.end < self.selected_range.start {
-            self.selection_reversed = !self.selection_reversed;
-            self.selected_range = self.selected_range.end..self.selected_range.start;
-        }
-        cx.notify()
-    }
-
-    fn offset_from_utf16(&self, offset: usize) -> usize {
-        let mut utf8_offset = 0;
-        let mut utf16_count = 0;
-
-        for ch in self.content.chars() {
-            if utf16_count >= offset {
-                break;
-            }
-            utf16_count += ch.len_utf16();
-            utf8_offset += ch.len_utf8();
-        }
-
-        utf8_offset
-    }
-
-    fn offset_to_utf16(&self, offset: usize) -> usize {
-        let mut utf16_offset = 0;
-        let mut utf8_count = 0;
-
-        for ch in self.content.chars() {
-            if utf8_count >= offset {
-                break;
-            }
-            utf8_count += ch.len_utf8();
-            utf16_offset += ch.len_utf16();
-        }
-
-        utf16_offset
-    }
-
-    fn range_to_utf16(&self, range: &Range<usize>) -> Range<usize> {
-        self.offset_to_utf16(range.start)..self.offset_to_utf16(range.end)
-    }
-
-    fn range_from_utf16(&self, range_utf16: &Range<usize>) -> Range<usize> {
-        self.offset_from_utf16(range_utf16.start)..self.offset_from_utf16(range_utf16.end)
-    }
-
-    fn previous_boundary(&self, offset: usize) -> usize {
-        self.content
-            .char_indices()
-            .rev()
-            .find_map(|(idx, _)| (idx < offset).then_some(idx))
-            .unwrap_or(0)
-    }
-
-    fn next_boundary(&self, offset: usize) -> usize {
-        self.content
-            .char_indices()
-            .find_map(|(idx, _)| (idx > offset).then_some(idx))
-            .unwrap_or(self.content.len())
-    }
-}
-
-impl EntityInputHandler for TextInput {
-    fn text_for_range(
-        &mut self,
-        range_utf16: Range<usize>,
-        actual_range: &mut Option<Range<usize>>,
-        _window: &mut Window,
-        _cx: &mut Context<Self>,
-    ) -> Option<String> {
-        let range = self.range_from_utf16(&range_utf16);
-        actual_range.replace(self.range_to_utf16(&range));
-        Some(self.content[range].to_string())
-    }
-
-    fn selected_text_range(
-        &mut self,
-        _ignore_disabled_input: bool,
-        _window: &mut Window,
-        _cx: &mut Context<Self>,
-    ) -> Option<UTF16Selection> {
-        Some(UTF16Selection {
-            range: self.range_to_utf16(&self.selected_range),
-            reversed: self.selection_reversed,
-        })
-    }
-
-    fn marked_text_range(
-        &self,
-        _window: &mut Window,
-        _cx: &mut Context<Self>,
-    ) -> Option<Range<usize>> {
-        self.marked_range
-            .as_ref()
-            .map(|range| self.range_to_utf16(range))
-    }
-
-    fn unmark_text(&mut self, _window: &mut Window, _cx: &mut Context<Self>) {
-        self.marked_range = None;
-    }
-
-    fn replace_text_in_range(
-        &mut self,
-        range_utf16: Option<Range<usize>>,
-        new_text: &str,
-        _: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let range = range_utf16
-            .as_ref()
-            .map(|range_utf16| self.range_from_utf16(range_utf16))
-            .or(self.marked_range.clone())
-            .unwrap_or(self.selected_range.clone());
-        let new_text = self.filter_input(&range, new_text);
-
-        self.content =
-            (self.content[0..range.start].to_owned() + &new_text + &self.content[range.end..])
-                .into();
-        self.selected_range = range.start + new_text.len()..range.start + new_text.len();
-        self.marked_range.take();
-        cx.notify();
-    }
-
-    fn replace_and_mark_text_in_range(
-        &mut self,
-        range_utf16: Option<Range<usize>>,
-        new_text: &str,
-        new_selected_range: Option<Range<usize>>,
-        _window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let range = range_utf16
-            .as_ref()
-            .map(|range_utf16| self.range_from_utf16(range_utf16))
-            .or(self.marked_range.clone())
-            .unwrap_or(self.selected_range.clone());
-        let new_text = self.filter_input(&range, new_text);
-
-        self.content =
-            (self.content[0..range.start].to_owned() + &new_text + &self.content[range.end..])
-                .into();
-        self.selected_range = new_selected_range
-            .as_ref()
-            .map(|range| self.range_from_utf16(range))
-            .unwrap_or(range.start + new_text.len()..range.start + new_text.len());
-        self.marked_range = new_selected_range.map(|range| self.range_from_utf16(&range));
-        cx.notify();
-    }
-
-    fn bounds_for_range(
-        &mut self,
-        range_utf16: Range<usize>,
-        bounds: Bounds<Pixels>,
-        _window: &mut Window,
-        _cx: &mut Context<Self>,
-    ) -> Option<Bounds<Pixels>> {
-        let last_layout = self.last_layout.as_ref()?;
-        let range = self.range_from_utf16(&range_utf16);
-        Some(Bounds::from_corners(
-            point(
-                bounds.left() + last_layout.x_for_index(range.start),
-                bounds.top(),
-            ),
-            point(
-                bounds.left() + last_layout.x_for_index(range.end),
-                bounds.bottom(),
-            ),
-        ))
-    }
-
-    fn character_index_for_point(
-        &mut self,
-        point: Point<Pixels>,
-        _window: &mut Window,
-        _cx: &mut Context<Self>,
-    ) -> Option<usize> {
-        let line_point = self.last_bounds?.localize(&point)?;
-        let last_layout = self.last_layout.as_ref()?;
-
-        let utf8_index = last_layout.index_for_x(point.x - line_point.x)?;
-        Some(self.offset_to_utf16(utf8_index))
-    }
-}
-
-struct TextElement {
-    input: Entity<TextInput>,
-}
-
-struct PrepaintState {
-    line: Option<ShapedLine>,
-    cursor: Option<PaintQuad>,
-    selection: Option<PaintQuad>,
-}
-
-fn snap_to_device(value: Pixels, scale: f32) -> Pixels {
-    let snapped = value.scale(scale).round();
-    px((f64::from(snapped) as f32) / scale)
-}
-
-impl IntoElement for TextElement {
-    type Element = Self;
-
-    fn into_element(self) -> Self::Element {
-        self
-    }
-}
-
-impl Element for TextElement {
-    type RequestLayoutState = ();
-    type PrepaintState = PrepaintState;
-
-    fn id(&self) -> Option<ElementId> {
-        None
-    }
-
-    fn source_location(&self) -> Option<&'static core::panic::Location<'static>> {
-        None
-    }
-
-    fn request_layout(
-        &mut self,
-        _id: Option<&GlobalElementId>,
-        _inspector_id: Option<&gpui::InspectorElementId>,
-        window: &mut Window,
-        cx: &mut App,
-    ) -> (LayoutId, Self::RequestLayoutState) {
-        let mut style = Style::default();
-        style.size.width = relative(1.).into();
-        style.size.height = window.line_height().into();
-        (window.request_layout(style, [], cx), ())
-    }
-
-    fn prepaint(
-        &mut self,
-        _id: Option<&GlobalElementId>,
-        _inspector_id: Option<&gpui::InspectorElementId>,
-        bounds: Bounds<Pixels>,
-        _request_layout: &mut Self::RequestLayoutState,
-        window: &mut Window,
-        cx: &mut App,
-    ) -> Self::PrepaintState {
-        let input = self.input.read(cx);
-        let content = input.content.clone();
-        let selected_range = input.selected_range.clone();
-        let cursor = input.cursor_offset();
-        let focused = input.focus_handle.is_focused(window);
-        let blink_on = if focused {
-            (input.blink_start.elapsed().as_millis() / CURSOR_BLINK_INTERVAL.as_millis())
-                .is_multiple_of(2)
-        } else {
-            false
-        };
-        let style = window.text_style();
-
-        let (display_text, text_color) = if content.is_empty() {
-            (input.placeholder.clone(), hsla(0., 0., 0.65, 1.0))
-        } else {
-            (content, style.color)
-        };
-
-        let run = TextRun {
-            len: display_text.len(),
-            font: style.font(),
-            color: text_color,
-            background_color: None,
-            underline: None,
-            strikethrough: None,
-        };
-        let runs = if let Some(marked_range) = input.marked_range.as_ref() {
-            vec![
-                TextRun {
-                    len: marked_range.start,
-                    ..run.clone()
-                },
-                TextRun {
-                    len: marked_range.end - marked_range.start,
-                    underline: Some(UnderlineStyle {
-                        color: Some(run.color),
-                        thickness: px(1.0),
-                        wavy: false,
-                    }),
-                    ..run.clone()
-                },
-                TextRun {
-                    len: display_text.len() - marked_range.end,
-                    ..run
-                },
-            ]
-            .into_iter()
-            .filter(|run| run.len > 0)
-            .collect()
-        } else {
-            vec![run]
-        };
-
-        let font_size = style.font_size.to_pixels(window.rem_size());
-        let line = window
-            .text_system()
-            .shape_line(display_text, font_size, &runs, None);
-
-        let cursor_pos = line.x_for_index(cursor);
-        let (selection, cursor) = if selected_range.is_empty() && blink_on {
-            let bounds_height: f32 = bounds.size.height.into();
-            let font_height: f32 = font_size.into();
-            let cursor_height = (font_height * 1.2).min(bounds_height);
-            let scale = window.scale_factor();
-            let cursor_top = snap_to_device(
-                bounds.top() + px((bounds_height - cursor_height) * 0.5),
-                scale,
-            );
-            let cursor_x = snap_to_device(bounds.left() + cursor_pos, scale);
-            let cursor_width = px(1.0 / scale);
-            (
-                None,
-                Some(fill(
-                    Bounds::new(
-                        point(cursor_x, cursor_top),
-                        size(cursor_width, px(cursor_height)),
-                    ),
-                    rgb(0xffffff),
-                )),
-            )
-        } else {
-            (
-                Some(fill(
-                    Bounds::from_corners(
-                        point(
-                            bounds.left() + line.x_for_index(selected_range.start),
-                            bounds.top(),
-                        ),
-                        point(
-                            bounds.left() + line.x_for_index(selected_range.end),
-                            bounds.bottom(),
-                        ),
-                    ),
-                    rgba(0x2d78ff55),
-                )),
-                None,
-            )
-        };
-        PrepaintState {
-            line: Some(line),
-            cursor,
-            selection,
-        }
-    }
-
-    fn paint(
-        &mut self,
-        _id: Option<&GlobalElementId>,
-        _inspector_id: Option<&gpui::InspectorElementId>,
-        bounds: Bounds<Pixels>,
-        _request_layout: &mut Self::RequestLayoutState,
-        prepaint: &mut Self::PrepaintState,
-        window: &mut Window,
-        cx: &mut App,
-    ) {
-        let focus_handle = self.input.read(cx).focus_handle.clone();
-        window.handle_input(
-            &focus_handle,
-            ElementInputHandler::new(bounds, self.input.clone()),
-            cx,
-        );
-        if let Some(selection) = prepaint.selection.take() {
-            window.paint_quad(selection)
-        }
-        let line = prepaint.line.take().unwrap();
-        line.paint(bounds.origin, window.line_height(), window, cx)
-            .unwrap();
-
-        let is_focused = focus_handle.is_focused(window);
-        if is_focused && let Some(cursor) = prepaint.cursor.take() {
-            window.paint_quad(cursor);
-        }
-        if is_focused {
-            window.request_animation_frame();
-        }
-
-        self.input.update(cx, |input, _cx| {
-            input.last_layout = Some(line);
-            input.last_bounds = Some(bounds);
-            if input.last_focused != is_focused {
-                input.last_focused = is_focused;
-                input.blink_start = Instant::now();
-            }
-        });
-    }
-}
-
-impl Render for TextInput {
-    fn render(&mut self, _window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
-        let border_color = if self.invalid {
-            hsla(0.0, 0.7, 0.55, 1.0)
-        } else {
-            rgb(0x2f2f2f).into()
-        };
-        div()
-            .flex()
-            .key_context("ConfigTextInput")
-            .track_focus(&self.focus_handle(cx))
-            .tab_stop(true)
-            .cursor(CursorStyle::IBeam)
-            .on_action(cx.listener(Self::backspace))
-            .on_action(cx.listener(Self::delete))
-            .on_action(cx.listener(Self::left))
-            .on_action(cx.listener(Self::right))
-            .on_action(cx.listener(Self::select_left))
-            .on_action(cx.listener(Self::select_right))
-            .on_action(cx.listener(Self::select_all))
-            .on_action(cx.listener(Self::home))
-            .on_action(cx.listener(Self::end))
-            .on_action(cx.listener(Self::tab))
-            .on_action(cx.listener(Self::tab_prev))
-            .on_action(cx.listener(Self::enter))
-            .on_action(cx.listener(Self::show_character_palette))
-            .on_action(cx.listener(Self::paste))
-            .on_action(cx.listener(Self::cut))
-            .on_action(cx.listener(Self::copy))
-            .on_mouse_down(MouseButton::Left, cx.listener(Self::on_mouse_down))
-            .on_mouse_up(MouseButton::Left, cx.listener(Self::on_mouse_up))
-            .on_mouse_up_out(MouseButton::Left, cx.listener(Self::on_mouse_up))
-            .on_mouse_move(cx.listener(Self::on_mouse_move))
-            .line_height(px(INPUT_HEIGHT))
-            .text_size(px(12.0))
-            .text_color(hsla(0.0, 0.0, 0.9, 1.0))
-            .child(
-                div()
-                    .flex()
-                    .items_center()
-                    .h(px(INPUT_HEIGHT))
-                    .w_full()
-                    .px(px(8.0))
-                    .py(px(0.0))
-                    .rounded(px(6.0))
-                    .bg(rgb(0x101010))
-                    .border_1()
-                    .border_color(border_color)
-                    .child(TextElement { input: cx.entity() }),
-            )
-    }
-}
-
-impl Focusable for TextInput {
-    fn focus_handle(&self, _: &App) -> FocusHandle {
-        self.focus_handle.clone()
-    }
-}
-
-pub fn bind_text_input_keys(cx: &mut App) {
-    cx.bind_keys([
-        KeyBinding::new("backspace", Backspace, None),
-        KeyBinding::new("delete", Delete, None),
-        KeyBinding::new("left", Left, None),
-        KeyBinding::new("right", Right, None),
-        KeyBinding::new("shift-left", SelectLeft, None),
-        KeyBinding::new("shift-right", SelectRight, None),
-        KeyBinding::new("cmd-a", SelectAll, None),
-        KeyBinding::new("cmd-v", Paste, None),
-        KeyBinding::new("cmd-c", Copy, None),
-        KeyBinding::new("cmd-x", Cut, None),
-        KeyBinding::new("home", Home, None),
-        KeyBinding::new("end", End, None),
-        KeyBinding::new("tab", Tab, None),
-        KeyBinding::new("shift-tab", TabPrev, None),
-        KeyBinding::new("enter", Enter, None),
-        KeyBinding::new("ctrl-cmd-space", ShowCharacterPalette, None),
-    ]);
 }
