+++ crates/subtitle-fast/src/gui/components/mod.rs
@@ -4,6 +4,7 @@ pub mod confirm_dialog;
 pub mod detection_sidebar;
 pub mod download_window;
 pub mod help_window;
+pub mod inputs;
 pub mod menu_bar;
 pub mod menu_bar_buttons;
 pub mod sidebar;
@@ -18,7 +19,7 @@ pub mod video_roi_overlay;
 pub mod video_toolbar;
 
 pub use color_picker::{ColorPicker, ColorPickerHandle};
-pub use config_editor::{ConfigWindow, bind_text_input_keys};
+pub use config_editor::ConfigWindow;
 pub use confirm_dialog::{
     ConfirmDialog, ConfirmDialogButton, ConfirmDialogButtonStyle, ConfirmDialogConfig,
     ConfirmDialogTitle,
@@ -29,6 +30,7 @@ pub use detection_sidebar::{
 };
 pub use download_window::DownloadWindow;
 pub use help_window::HelpWindow;
+pub use inputs::bind_text_input_keys;
 pub use menu_bar::MenuBar;
 pub use sidebar::{
     CollapseDirection, DragRange, DraggableEdge, Sidebar, SidebarConfig, SidebarHandle,
